feat: fix memory layering mechanism and add conflict resolution UI
- Fix layer routing: new memories default to flash/short instead of long, tighten KEEP_LONG_KEYWORDS to title-only matching

- Add real decay curves per layer (long 180d / short 30d / flash 7d)

- Add WorkingSetView with three columns: recent hits, recent writes, loop runs

- Backfill project column on write via JOIN, unify timestamp format to ISO

- Add conflict resolution: detectConflictMemories stores targetId, new resolveConflict() supports keep/merge/delete, new POST /api/dream/conflicts/resolve endpoint

- Add ConflictResolveDialog with merge/keep/delete modes, side-by-side memory comparison, editable merged content

- Update i18n with 21 conflict resolution strings (zh/en)

+++ packages/server/src/core/migration.ts
@@ -376,7 +376,9 @@ function normalizeRawMemory(raw: RawMemory, options: MigrationOptions, index: nu
   return {
     title,
     content,
-    layer: validLayer(raw.layer) ?? options.defaultLayer ?? (memoryKind === 'task' ? 'short' : 'long'),
+    // 迁移默认层从 long 改为 short：避免历史长内容一律进 long 形成只进不出
+    // task 类→short；其余未指定→short，由 dream 升格到 long
+    layer: validLayer(raw.layer) ?? options.defaultLayer ?? (memoryKind === 'task' ? 'short' : 'short'),
     projectPath,
     tags,
     metadata: {
